modularize the code structure of storyBook
