+++ src/storybook/stories/Decorators/EmptyNavigator/index.js
@@ -1,0 +1,11 @@
+import { createAppContainer, createSwitchNavigator } from 'react-navigation';
+import React from 'react';
+
+const reactNavigationDecorator = story => {
+	const Screen = () => story();
+	const Navigator = createAppContainer(createSwitchNavigator({ Screen }));
+	return <Navigator />;
+};
+
+export default reactNavigationDecorator;
+
