+++ src/storybook/stories/WelcomeView/Welcome/index.js
@@ -34,14 +34,7 @@ export default class Welcome extends React.Component {
 			<View style={this.styles.wrapper}>
 				<Text style={this.styles.header}>Welcome to React Native Storybook</Text>
 				<Text style={this.styles.content}>
-          This is a UI Component development environment for your React Native app. Here you can
-          display and interact with your UI components as stories. A story is a single state of one
-          or more UI components. You can have as many stories as you want. In other words a story is
-          like a visual test case.
-				</Text>
-				<Text style={this.styles.content}>
-          We have added some stories inside the "storybook/stories" directory for examples. Try
-          editing the "storybook/stories/Welcome.js" file to edit this message.
+                  Hi, here is the storybook for banana App, you could see all the documented components in navigator, and the related documentation is on wiki of our github repo. Visit wiki of any component via link attached in add-on notes.
 				</Text>
 			</View>
 		);
